requirements.txt
Django==3.0.7
django-rest-auth==0.9.5
djangorestframework==3.11.0
requests==2.24.0
